requirements.txt
feedparser==6.0.12
requests==2.32.5
schedule==1.2.2
dashscope==1.25.12
python-dotenv==1.2.2
protobuf==5.29.6
Markdown==3.10.2

